+++ test/prost/core_test.cljs
@@ -3,14 +3,20 @@
             [prost.core :refer-macros [arg! ret! shape!]]
             [cljs.spec.alpha :as s]))
 
+(def foo-bar-set #{:foo :bar})
+
 (s/def ::pos-int (s/and integer? pos?))
 (s/def ::fooish (s/keys :req-un [::pos-int]))
+(s/def ::contains foo-bar-set)
+(s/def ::contains-key (s/keys :req-un [::contains]))
 
 (deftest arg!
-  (let [expect-string  (fn [v] {:pre [(arg! string? v)]} [:ret v])
+  (let [expect-string (fn [v] {:pre [(arg! string? v)]} [:ret v])
         expect-integer (fn [v] {:pre [(arg! integer? v)]} [:ret v])
         expect-pos-int (fn [v] {:pre [(arg! ::pos-int v)]} [:ret v])
-        expect-fooish (fn [v] {:pre [(arg! ::fooish v)]} [:ret v])]
+        expect-fooish (fn [v] {:pre [(arg! ::fooish v)]} [:ret v])
+        ; expect-contains (fn [v] {:pre [(arg! ::contains v)]} [:ret v])
+        expect-contains-key (fn [v] {:pre [(arg! ::contains-key v)]} [:ret v])]
 
     (testing "with valid predicate fn"
       (is (expect-string "asdf"))
@@ -34,13 +40,20 @@
                             (expect-pos-int -12)))
       (is (thrown-with-msg? js/TypeError
                             #"invalid argument \'v \:pos\-int\', expected \-12 to be pos\? via \:prost\.core\-test/fooish > \:prost\.core\-test/pos\-in"
-                            (expect-fooish {:pos-int -12}))))))
+                            (expect-fooish {:pos-int -12})))
+      ; (is (thrown-with-msg? js/TypeError
+      ;                       #"invalid argument \'v\', expected :baz to be one of the allowed values #\{:bar :foo} via \:prost\.core\-test/contains"
+      ;                       (expect-contains :baz)))
+      (is (thrown-with-msg? js/TypeError
+                            #"invalid argument \'v\', expected \{:foo :bar} to contain the key :contains via \:prost\.core\-test/contains-key"
+                            (expect-contains-key {:foo :bar}))))))
 
 (deftest ret!
   (let [expect-string  (fn [v] {:post [(ret! string? %)]} v)
         expect-integer (fn [v] {:post [(ret! integer? %)]} v)
         expect-pos-int (fn [v] {:post [(ret! ::pos-int %)]} v)
-        expect-fooish (fn [v] {:post [(ret! ::fooish %)]} v)]
+        expect-fooish (fn [v] {:post [(ret! ::fooish %)]} v)
+        expect-contains-key (fn [v] {:pre [(ret! ::contains-key v)]} v)]
     (testing "with valid predicate fn"
       (is (expect-string "asdf"))
       (is (expect-integer 12)))
@@ -63,7 +76,10 @@
                             (expect-pos-int -12)))
       (is (thrown-with-msg? js/TypeError
                             #"invalid return value \':pos-int\', expected \-12 to be pos\? via \:prost\.core\-test/fooish > \:prost\.core\-test/pos\-in"
-                            (expect-fooish {:pos-int -12}))))))
+                            (expect-fooish {:pos-int -12})))
+      (is (thrown-with-msg? js/TypeError
+                            #"invalid return value, expected \{:foo :bar} to contain the key :contains via \:prost\.core\-test/contains-key"
+                            (expect-contains-key {:foo :bar}))))))
 
 (deftest shape!
   (let [expect-fooish (fn [v] (shape! "foo" ::fooish v))]
@@ -72,4 +88,7 @@
     (testing "with invalid key"
       (is (thrown-with-msg? js/TypeError
                             #"invalid shape \'foo \:pos\-int\', expected \-12 to be pos\? via \:prost\.core\-test/fooish > \:prost\.core\-test/pos\-in"
-                            (expect-fooish {:pos-int -12}))))))
+                            (expect-fooish {:pos-int -12})))
+      (is (thrown-with-msg? js/TypeError
+                            #"invalid shape \'foo\', expected \{:foo :bar} to contain the key :pos-int via \:prost\.core\-test/fooish"
+                            (expect-fooish {:foo :bar}))))))